move table to home screen
